my-app/api/requirements.txt
fastapi
pydantic
pandas
beautifulsoup4
html2text
tiktoken
python-dotenv
selenium
Pillow
pytesseract
2captcha-python
openai
chromedriver-autoinstaller
requests
uvicorn
